requirements.txt

Flask
flask-cors
python-dotenv
requests
